search-engine: more cleanup and simplification
Fixed a typo in not-yet-used code, added credit for `blocked`, and
changed the way the skip file is built so that it’s built in the
correct order, doesn’t require a second pass, avoids a special case,

Also simplified merge output name generation and fixed a bug for terms
that come earlier than any headword.

This is going to require some updates to the chapter text.