Make the api for getting the defined function nicer
Rather than a get_function(name) method, we can simply ask for all the
globals defined by the code, and then get the one we want from the
dictionary.  This removes the need to pass in the function name,
and means that we could some day use this class to define more than one
function.