updating search-engine in response to feedback
This incorporates some of the reviewers' feedback.  Also this is the
first version of index.py that is built from README.md, and therefore
arguably it shouldn't be checked in at all, but checking it in enables
other people to run the search engine without rebuilding it from the
document.  Like a `configure` script.