ignore this suggestion
I've gone back and forth on this, trying to find a good way to explain
the needed refactoring. The problem with it is that it just takes too
explanation a la "now we change the class in this non-trivial way" which
feels to me a lot less natural than the current slightly clumsy extra
class.